fix: fix error updating project description
When you edit the project description on `EditProject` screen,
it doesn't update in the store.
This is because of an error in the function that validates and saves the changes.

The function is wrapped with a `useCallback` hook, but the input state for the project description was missing in its dependency array.